Update - refactor testcases
