+++ agenta-backend/agenta_backend/tests/conftest.py
@@ -3,8 +3,6 @@ import asyncio
 
 from agenta_backend.models.db_engine import DBEngine
 
-engine = DBEngine().engine()
-
 
 @pytest.fixture(scope="session", autouse=True)
 def event_loop():
@@ -14,7 +12,10 @@ def event_loop():
     asyncio.set_event_loop(res)
     res._close = res.close
 
+    # Initialize beanie
+    res.run_until_complete(DBEngine().init_db())
+
     yield res
 
     res._close()  # close event loop
-    DBEngine().remove_db()  # drop database
+    # DBEngine().remove_db()  # drop database
