+++ agenta-backend/agenta_backend/tests/variants_main_router/test_variant_evaluators_router.py
@@ -6,7 +6,9 @@ import asyncio
 from agenta_backend.models.api.evaluation_model import EvaluationStatusEnum
 from agenta_backend.models.db_models import (
     AppDB,
+    ConfigDB,
     TestSetDB,
+    AppVariantDB,
     EvaluationDB,
     AppVariantDB,
     DeploymentDB,
@@ -146,6 +148,25 @@ async def test_get_evaluator_configs():
 
 
 @pytest.mark.asyncio
+async def test_update_app_variant_parameters(update_app_variant_parameters):
+    app = await AppDB.find_one(AppDB.app_name == APP_NAME)
+    testset = await TestSetDB.find_one(TestSetDB.app.id == app.id)
+    app_variant = await AppVariantDB.find_one(
+        AppVariantDB.app.id == app.id,
+        AppVariantDB.variant_name == "app.default"
+    )
+
+    parameters = update_app_variant_parameters
+    parameters["inputs"] = [{"name": list(testset.csvdata[0].keys())[0]}]
+    payload = {"parameters": parameters}
+
+    response = await test_client.put(
+        f"{BACKEND_API_HOST}/variants/{str(app_variant.id)}/parameters/", json=payload
+    )
+    assert response.status_code == 200
+
+
+@pytest.mark.asyncio
 async def test_create_evaluation():
     # Fetch app, app_variant and testset
     app = await AppDB.find_one(AppDB.app_name == APP_NAME)
@@ -193,7 +214,9 @@ async def test_create_evaluation():
 
 @pytest.mark.asyncio
 async def test_fetch_evaluation_status():
-    evaluations = await EvaluationDB.find().to_list()  # will return only one in this case
+    evaluations = (
+        await EvaluationDB.find().to_list()
+    )  # will return only one in this case
     evaluation = evaluations[0]
 
     # Prepare and start short-polling request
@@ -217,7 +240,9 @@ async def test_fetch_evaluation_status():
 
 @pytest.mark.asyncio
 async def test_fetch_evaluation_results():
-    evaluations = await EvaluationDB.find().to_list()  # will return only one in this case
+    evaluations = (
+        await EvaluationDB.find().to_list()
+    )  # will return only one in this case
     evaluation = evaluations[0]
 
     response = await test_client.get(
@@ -252,9 +277,13 @@ async def test_delete_evaluator_config():
 
 @pytest.mark.asyncio
 async def test_evaluation_scenario_match_evaluation_testset_length():
-    evaluations = await EvaluationDB.find(fetch_links=True).to_list()  # will return only one in this case
+    evaluations = await EvaluationDB.find(
+        fetch_links=True
+    ).to_list()  # will return only one in this case
     evaluation = evaluations[0]
-    evaluation_scenario_count = await EvaluationScenarioDB.find(EvaluationScenarioDB.evaluation.id == evaluation.id).count()
+    evaluation_scenario_count = await EvaluationScenarioDB.find(
+        EvaluationScenarioDB.evaluation.id == evaluation.id
+    ).count()
 
     assert evaluation_scenario_count == len(evaluation.testset.csvdata)
