+++ agenta-backend/agenta_backend/tests/variants_main_router/test_variant_testset_router.py
@@ -28,7 +28,7 @@ elif ENVIRONMENT == "github":
 
 @pytest.mark.asyncio
 async def test_create_testset():
-    app = await AppDB.find_one(AppDB.app_name == "test_app")
+    app = await AppDB.find_one(AppDB.app_name == "app_variant_test")
 
     payload = {
         "name": "create_testset_main",
@@ -50,15 +50,14 @@ async def test_create_testset():
     response = await test_client.post(
         f"{BACKEND_API_HOST}/testsets/{str(app.id)}/", json=payload
     )
-
     assert response.status_code == 200
     assert response.json()["name"] == payload["name"]
 
 
 @pytest.mark.asyncio
 async def test_update_testset():
-    app = await AppDB.find_one(AppDB.app_name == "test_app")
-    testset = await AppDB.find_one(TestSetDB.app.id == app.id)
+    app = await AppDB.find_one(AppDB.app_name == "app_variant_test")
+    testset = await TestSetDB.find_one(TestSetDB.app.id == app.id)
 
     payload = {
         "name": "update_testset",
@@ -89,7 +88,7 @@ async def test_update_testset():
 
 @pytest.mark.asyncio
 async def test_get_testsets():
-    app = await AppDB.find_one(AppDB.app_name == "test_app")
+    app = await AppDB.find_one(AppDB.app_name == "app_variant_test")
     response = await test_client.get(
         f"{BACKEND_API_HOST}/testsets/?app_id={str(app.id)}"
     )
@@ -100,7 +99,7 @@ async def test_get_testsets():
 
 @pytest.mark.asyncio()
 async def test_get_testset():
-    app = await AppDB.find_one(AppDB.app_name == "test_app")
+    app = await AppDB.find_one(AppDB.app_name == "app_variant_test")
     testset = await TestSetDB.find_one(TestSetDB.app.id == app.id)
 
     response = await test_client.get(f"{BACKEND_API_HOST}/testsets/{str(testset.id)}/")
@@ -112,8 +111,8 @@ async def test_get_testset():
 
 @pytest.mark.asyncio
 async def test_delete_testsets():
-    app = await AppDB.find_one(AppDB.app_name == "test_app")
-    testsets = await TestSetDB.find(TestSetDB.app.id == app.id)
+    app = await AppDB.find_one(AppDB.app_name == "app_variant_test")
+    testsets = await TestSetDB.find(TestSetDB.app.id == app.id).to_list()
 
     testset_ids = [str(testset.id) for testset in testsets]
     payload = {"testset_ids": testset_ids}
